ensuring modal unmounting for the account settings when changing the account
